+++ Changelog.md
@@ -8,6 +8,8 @@ Compiler Features:
 
 Bugfixes:
  * SMTChecker: Fix internal error when using bitwise operators with an array element as argument.
+ * Commandline Interface: Fix ICE when the optimizer is disabled and an empty/blank string is used for ``--yul-optimizations`` sequence.
+ * Standard JSON Interface: Fix ICE when the optimizer is disabled and an empty/blank string is used for ``optimizerSteps`` sequence.
 
 
 ### 0.8.25 (2023-03-14)